move `merge_eopatches` back to `eodata_merge`
